api/requirements.txt
pytest==6.2.2
marshmallow==3.10.0
Flask==1.1.2
fakeredis==1.4.5
Werkzeug==1.0.1
redis==3.5.3
PyJWT==2.0.1
python-dotenv==0.15.0